requirements.txt
psycopg2==2.9.10
python-dotenv==1.0.1
Unidecode==1.3.8
